updated sauvegarde (praticly finished)

+++ src/main/java/Gaufre/Modele/Coup.java
@@ -21,10 +21,12 @@ public class Coup {
 
     public Coup(String sauvegarde) {
         sauvegarde = sauvegarde.substring(1, sauvegarde.length() - 1);
+
         String parties[] = sauvegarde.split(";");
 
         String point[] = parties[0].substring(1,parties[0].length() - 1).split(",");
-        position = new Point(Integer.parseInt(point[0]), Integer.parseInt(point[1]));
+
+        position = new Point((int) Integer.parseInt(point[0]), (int) Integer.parseInt(point[1]));
 
         String plateau[] = parties[1].substring(1, parties[1].length() - 1).split(",");
         int taille = plateau.length;
@@ -60,7 +62,7 @@ public class Coup {
 
     public String pourSauvegarde() {
         String s = "{";
-        s += "(" + getPosition().getX() + "," + getPosition().getY() + ");";
+        s += "(" + (int) getPosition().getX() + "," + (int) getPosition().getY() + ");";
         String plateau = "";
         for (int i = 0; i < getAncienPlateau().length; i++) {
             plateau += getAncienPlateau()[i];
